feat(dsg):resolver base
